Fixed module name

+++ basics/switch.go
@@ -11,7 +11,7 @@ func SwitchTest(v interface{}) {
 		fmt.Println("Value received is of type string")
 	case int:
 		fmt.Println("Value received is of type int")
-	case float32:
+	case float64:
 		fmt.Println("Value received is of type float32")
 	default:
 		fmt.Println("Value type unknown")
